+++ packages/mcp/CHANGELOG.md
@@ -1,7 +1,11 @@
 # CHANGELOG
 
+## 0.1.4 - Unreleased
+    
+    - Added MemoryServer to use LightRAG
+
 ## 0.1.3 - Jan 14 2025 5:00 PM
 
-- Modified `registerCharacter` tool to return pubkey if it was already registered
+    - Modified `registerCharacter` tool to return pubkey if it was already registered
 
 ## 0.1.2 - Initial Release